chore (refs T22614): move News Components to client/js.
