+++ client/js/components/news/DpNewsAdminList.vue
