+++ client/js/components/news/DpNewsItemStatus.vue
